+++ R/track_sensor.R
@@ -244,8 +244,7 @@ track_sensor.LAScatalog <- function(las, algorithm, extra_check = TRUE, thin_pul
   }
 
   options <- list(need_buffer = TRUE, drop_null = TRUE, need_output_file = FALSE)
-  output  <- catalog_apply(las, track_sensor, algorithm = algorithm, extra_check = extra_check, thin_pulse_with_time = 0, multi_pulse = multi_pulse, .options = options)
-  output  <- do.call(rbind, output)
+  output  <- catalog_sapply(las, track_sensor, algorithm = algorithm, extra_check = extra_check, thin_pulse_with_time = 0, multi_pulse = multi_pulse, .options = options)
 
   # Post-processing to remove duplicated positions selecting the one computed with the more pulses
   SCORE <- gpstime <- NULL
@@ -254,5 +253,7 @@ track_sensor.LAScatalog <- function(las, algorithm, extra_check = TRUE, thin_pul
   i <- data[, .I[which.max(SCORE)], by = gpstime]$V1
   data.table::setDF(data)
 
-  return(output[i,])
+  ans <- output[i,]
+  if (nrow(ans) == 0)  warning("0 sensor locations found. This may be caused by an unsufficient number of multiple returns.", call. = FALSE)
+  return(ans)
 }